feat: move folder loading from `common` into `loading`
