+++ src/plugins/common/src/systems.rs
@@ -5,9 +5,7 @@ pub mod log;
 pub mod remove_components;
 pub mod track_components;
 
-pub(crate) mod begin_loading_folder_assets;
 pub(crate) mod load_asset_model;
-pub(crate) mod map_load_results;
 pub(crate) mod remove_not_owned;
 pub(crate) mod set_cam_ray;
 pub(crate) mod set_mouse_hover;
