+++ src/plugins/common/src/traits.rs
@@ -1,7 +1,5 @@
 pub mod accessors;
 pub mod add_asset;
-pub mod asset_file_extensions;
-pub mod asset_folder;
 pub mod cache;
 pub mod cast_ray;
 pub mod clamp_zero_positive;
@@ -18,7 +16,6 @@ pub mod iterate;
 pub mod iteration;
 pub mod load_asset;
 pub mod load_folder_assets;
-pub mod load_from;
 pub mod map_value;
 pub mod mock;
 pub mod mouse_position;
@@ -26,7 +23,6 @@ pub mod nested_mock;
 pub mod ownership_relation;
 pub mod process_delta;
 pub mod read;
-pub mod register_custom_folder_assets;
 pub mod remove_conditionally;
 pub mod remove_resource;
 pub mod state_duration;
