+++ src/plugins/loading/Cargo.toml
@@ -5,6 +5,9 @@ edition.workspace = true
 
 [dependencies]
 bevy.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+uuid.workspace = true
 common.workspace = true
 macros.workspace = true
