+++ src/plugins/loading/src/systems.rs
@@ -1,1 +1,4 @@
 pub mod is_processing;
+
+pub(crate) mod begin_loading_folder_assets;
+pub(crate) mod map_load_results;
