+++ src/plugins/loading/src/traits.rs
@@ -1,2 +1,6 @@
+pub mod asset_file_extensions;
+pub mod asset_folder;
+pub mod load_from;
 pub mod progress;
+pub mod register_custom_folder_assets;
 pub mod register_load_tracking;
