+++ src/plugins/loading/src/traits/asset_file_extensions.rs
