+++ src/plugins/loading/src/traits/asset_folder.rs
@@ -1,4 +1,4 @@
-use super::load_asset::Path;
+use common::traits::load_asset::Path;
 
 pub trait AssetFolderPath {
 	fn asset_folder_path() -> Path;
