+++ src/plugins/loading/src/traits/load_from.rs
@@ -1,4 +1,4 @@
-use super::load_asset::LoadAsset;
+use common::traits::load_asset::LoadAsset;
 
 pub trait LoadFrom<TFrom> {
 	fn load_from<TLoadAsset: LoadAsset>(from: TFrom, asset_server: &mut TLoadAsset) -> Self;
