+++ src/plugins/loading/src/traits/register_custom_folder_assets.rs
