+++ src/plugins/skills/Cargo.toml
@@ -14,6 +14,7 @@ behaviors.workspace = true
 common.workspace = true
 interactions.workspace = true
 items.workspace = true
+loading.workspace = true
 player.workspace = true
 shaders.workspace = true
